Use grunt-angular-templates but disable grunt-rev
This change prevents each view change from making an ajax request which
speeds the app up a lot! The problem is the URLs for images in the
templates are not being fingerprinted by grunt-rev because the templates
are no longer html files, but stored in a js file.

Option #1:

Usemin can be configured to replace urls with the revved versions based
on configured patterns:
  https://github.com/yeoman/grunt-usemin#patterns

I did have some limited success adding the js files to its configuration,
but the patterns only ever seemed to apply to one asset at a time, I
couldn't get the pattern to apply to all at once and I don't want to
maintain one pattern per asset!

Option #2:

This other grunt plugin may be the answer:
https://github.com/wmluke/grunt-inline-angular-templates

This one will put the templates in a script tag in the html.
I also had some limited success with this. grunt-rev seems to
fingerprint the image urls, etc. but they don't seem to be loading into
Angular's template cache. I guess the plugin doesn't do it for you and I
havent really tried to do it myself.