Serve the header font ourselves
This improves page speed because we don't have to load Google's CSS
sheet first, then download the font from them.

I opted to serve the font as well because we are not getting a new
version of the CSS everytime, there is no guarantee the Google served
font will always match out potentially out of date, cached version of
its matching CSS file.