Extract a Player factory and store scoring history
Instead of maintaining the current score, it appends a scoreEvent to the
players scoreEvents history.
Then when asking a player for its score, it will iterate over the
history and calculate the current score (each time).
When a new game is started, this history is wiped so it doesn't continue
to grow forever.
This event sourcing model should make an 'undo' feature (#6) very easy!