Implement 'undo'!
Details in issue #6

Also, if a player score goes over the maximum is equal to (target-1),
the last score is poped off before pushing a '0' (to represent a foul).
This is because an undo after a foul would only pop the '0' off and
leave the player with an invalid score.