Add all the favicons and make behave like an app
The original favicon is still used for favicon.ico, favicon-16x16 and
favicon-32x32, but the poo logo from the from page is now used for all
the others as they are higher resolution anyway.

The manifest is also configured to allow Chrome on Android to hide the
navbar, etc. to make it behave like an app when saved to the homescreen.
As this app is offline-enabled and now has large favicons to support it,
it is very much like a native app!