A button to load the previous game
If someone was to accidentally click the header during the middle of a
game, there is no way to go back to that game except for the browser
back button. On an iPhone there is no physical back button and the
change in 15f33f4 removes the browser navbar which leaves no way of
getting back to the game.

A possibly better solution sould be to prevent or warn the user when
about to leave a game half way through with a solution like this:
https://stackoverflow.com/questions/1119289/how-to-show-the-are-you-sure-you-want-to-navigate-away-from-this-page-when-ch